feat(tui): ground-up rebuild -- intelligence pipeline, modular architecture, UI polish
Phase 1 -- Stable Foundation:
- Delete dead agents/ module (6 files, -620 lines)
- Extract 919-line main.rs into event_loop, dispatch, commands modules
- Wire ConversationManager as single source of truth for messages
- Rewrite sidecar lifecycle: PID file tracking, targeted kill, health backoff
- Graceful degradation for all external services (sidecar, daemon, Engram)

Phase 2 -- Intelligence Layer:
- Context compression via local LLM with caching (skip re-summarize)
- Prompt distillation: rambling intent -> structured task spec (GBNF constrained)
- Token budget estimation by model family (Qwen/Claude/OpenAI)
- Adaptive model selection based on complexity + budget
- Full pipeline: compress -> Engram search -> distill -> budget -> select
- User approval before dispatch with editable distilled prompt
- DatasetCollector extended with intelligence fields for fine-tuning

Phase 3 -- Response Synthesis and Polish:
- Response synthesizer: verbose agent output -> structured summary via local LLM
- Daemon prompt enrichment via generate_prompt() before submit_task()
- Multi-line input (Shift+Enter), dynamic input bar height
- scroll_offset u16 -> usize, Page Up/Down support
- Ctrl+E raw/synthesized output toggle, Ctrl+A agent panel toggle
- Zero warnings, 52 tests passing

+++ eidolon-tui/src/tui/widgets/chat_area.rs
@@ -17,11 +17,11 @@ pub struct ChatMessage {
 pub struct ChatArea<'a> {
     theme: &'a Theme,
     messages: &'a [ChatMessage],
-    scroll_offset: u16,
+    scroll_offset: usize,
 }
 
 impl<'a> ChatArea<'a> {
-    pub fn new(theme: &'a Theme, messages: &'a [ChatMessage], scroll_offset: u16) -> Self {
+    pub fn new(theme: &'a Theme, messages: &'a [ChatMessage], scroll_offset: usize) -> Self {
         Self { theme, messages, scroll_offset }
     }
 }
@@ -61,7 +61,7 @@ impl Widget for ChatArea<'_> {
 
         let paragraph = Paragraph::new(lines)
             .wrap(Wrap { trim: false })
-            .scroll((self.scroll_offset, 0));
+            .scroll((self.scroll_offset as u16, 0));
 
         paragraph.render(inner, buf);
     }
